# backend/docs/booking-flow.openapi.yaml
openapi: 3.1.0
info:
  title: UMC Web Booking Flow API
  version: 1.0.0
  description: >-
    API contract for redesigned patient web booking flow. Supports 3 entry modes
    (by date, by department, by doctor) and converges into unified booking + payment pipeline.
servers:
  - url: https://api.example.com
    description: Production
  - url: https://staging-api.example.com
    description: Staging
...
paths:
...
  /v1/insurance/bhyt-types:
    get:
      tags: [Insurance]
      summary: Get BHYT options
      parameters:
        - in: query
          name: profileId
          required: true
          schema:
            type: string
        - in: query
          name: departmentId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/BHYTType'
...
components:
...
  schemas:
...
    BHYTType:
      type: object
      required: [id, label]
      properties:
        id:
          type: string
        label:
          type: string
        description:
          type: string
          nullable: true
...